Check for zero money on faction disband
The takeOnDisband method of EngineEcon now checks if the amount of money
to be transferred to the player upon disbanding a faction is 0. If this is
the case, the method exits.

This will prevent unnecessary operations in transferring money to players
(which could be expensive if the economy is hooked to a database, for
example).